task 1, more bits. task 2, long int

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int i = n ^ m;
+	unsigned long int i = n ^ m;
 	int count = 0;
 
 	while (i)
